OGM-855 Rename package util to utils
