+++ orientdb/src/test/java/org/hibernate/datastore/ogm/orientdb/utils/OrientDBTestHelper.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.datastore.ogm.orientdb.util;
+package org.hibernate.datastore.ogm.orientdb.utils;
 
 import java.util.HashMap;
 import java.util.List;
